requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
tkinterdnd2>=0.3.0
pyinstaller>=6.0.0
onnxruntime>=1.16.0
